Abstract version1
