--- src/myn/patterns/decorator/GraphicRedactorDecorator/IShapeRedactor.java
@@ -1,5 +1,0 @@
-package myn.patterns.decorator.GraphicRedactorDecorator;
-
-public interface IShapeRedactor {
-    void draw(IShapeRedactor shape);
-}
