+++ src/myn/patterns/decorator/GraphicRedactorDecorator/Main.java
@@ -2,7 +2,8 @@ package myn.patterns.decorator.GraphicRedactorDecorator;
 
 public class Main {
     public static void main(String[] args) {
-        AbstractShapeRedactor shape = new MoveDownShape(new MoveUpShape(new ReduceShape(new IncreaseShape(new ShapeRedactor(3,3,10)))));
-        shape.draw(shape);
+        AbstractShapeRedactor shapeRedactor = new ShapeRedactor(3,3,10);
+        AbstractShapeRedactor shape = new MoveDownShape(new MoveUpShape(new ReduceShape(new IncreaseShape(shapeRedactor))));
+        shape.draw(shapeRedactor);
     }
 }
